Able to change lighting settings and other things (read description)
For testing and that can be used I've made two camera one with the forest lighting and another with the dark world lighting, you can switch between the two with 1 and 2. Bug that comes is that switching two 2 keeps the controls of 1 so you can't really move correctly. but I imagine with a screen transisction you can switch Camera for the dark world when on the whale. Added "thinking" text so when you do things the player thinks to them-self of what to do next or what they did (Added Template in game manager code for Julia to write what the character wants to think)