Changed 'thinking text', added bell swing (check description)
Hullo! I changed most of the thinking text except for the bottom two coz it didn't specify what it was for. I also placed the new wendigo in the dark world, he is just looping through is anims coz I don't wanna mess anything up that Ben has put in place. Also also, I added a simple anim for the bell swinging. It can be sped up or slowed down to match the sound effect. I am not sure how to make it play only when the cut scene runs, so atm it's just looping